requirements.txt
faker==30.1.0
pandas==2.2.2
python-dateutil==2.9.0.post0
kafka-python-ng==2.2.3
